label: use a C99 bool for 'label_params_t.force'

+++ lib/common/postproc.c
@@ -216,7 +216,7 @@ static void printData(object_t *objs, size_t n_objs, xlabel_t *lbls,
   xlabel_t* xp;
   fprintf (stderr, "%" PRISIZE_T " objs %" PRISIZE_T
 	   " xlabels force=%d bb=(%.02f,%.02f) (%.02f,%.02f)\n",
-	   n_objs, n_lbls, params->force, params->bb.LL.x, params->bb.LL.y,
+	   n_objs, n_lbls, (int)params->force, params->bb.LL.x, params->bb.LL.y,
 	   params->bb.UR.x, params->bb.UR.y);
   if (Verbose < 2) return;
   fprintf(stderr, "objects\n");
